refactor: decompose @blecsd/media files to meet size limits
- gif/parser.ts (842→364): split into types, helpers, frameRendering, main parser
- render/ansi.ts (744→122): split into types, helpers, modes, main renderer
- overlay/w3m.ts (738→165): split into types, helpers, commands, main module
- png/pixels.ts (511→326): extract helpers for palette/bit-depth operations
- widgets/video/factory.ts (431→324): extract playback helpers
- widgets/image/factory.ts (418→313): extract animation helpers

All files now under size limits (500 general, 300 widgets). No API changes.
